with flout

+++ src/floutgipotenuza.java
@@ -1,9 +1,9 @@
 public class floutgipotenuza {
     public static void main(String[] args) {
-        double a = 3.5;
-        double b = 4.8;
-        double c = Math.sqrt(a*a+b*b);
-        double sqrt = Math.sqrt(c);
+        float a = 3.5f;
+        float b = 4.8f;
+        float c = (float) Math.sqrt(a*a+b*b);
+        float sqrt = (float) Math.sqrt(c);
           System.out.println("Корен с равен"+sqrt);
     }
 }
